+++ src/main/java/org/pollub/campusmate/service/GradeService.java
@@ -2,6 +2,7 @@ package org.pollub.campusmate.service;
 
 import lombok.AllArgsConstructor;
 import org.pollub.campusmate.entity.Grade;
+import org.pollub.campusmate.exception.GradeNotFound;
 import org.pollub.campusmate.repository.GradeRepository;
 import org.springframework.stereotype.Service;
 
@@ -16,7 +17,9 @@ public class GradeService {
     private final GradeRepository gradeRepository;
 
     public Grade getGrade(long gradeId) {
-        return gradeRepository.findById(gradeId).orElseThrow(null);
+
+        return gradeRepository.findById(gradeId)
+                .orElseThrow(() -> new GradeNotFound("Grade with id " + gradeId + " not found"));
     }
 
     public void addGrade(Grade grade) {
@@ -24,9 +27,9 @@ public class GradeService {
     }
 
     public void deleteGrade(long gradeId) {
-//        if(gradeRepository.existsById(gradeId)) {
-//
-//        }
+        if(!gradeRepository.existsById(gradeId)) {
+            throw new GradeNotFound("Cannot execute delete operation. Grade with id " + gradeId + " not found");
+        }
         gradeRepository.deleteById(gradeId);
     }
 
@@ -37,7 +40,7 @@ public class GradeService {
         grades.forEach(foundGrades::add);
 
         if(foundGrades.isEmpty()){
-            return null;
+            throw new GradeNotFound("Grades not found");
         }
 
         return foundGrades;
